# docs/swagger/swagger.yaml
basePath: /
definitions:
  alert.AlertManagerAlert:
    properties:
      annotations:
        additionalProperties:
          type: string
        type: object
      endsAt:
        type: string
      fingerprint:
        type: string
      generatorURL:
        type: string
      labels:
        additionalProperties:
          type: string
        type: object
      startsAt:
        type: string
      status:
        type: string
    type: object
...
  model.AlertSubscriptionNode:
    properties:
      children:
        items:
          $ref: '#/definitions/model.AlertSubscriptionNode'
        type: array
      code:
        type: string
      continue:
        description: 路由行为
        type: boolean
      created_at:
        type: string
      enabled:
        type: boolean
      id:
        type: integer
      level:
        type: integer
      match_labels:
        additionalProperties:
          type: string
        description: 非数据库字段
        type: object
      match_labels_json:
        description: 匹配条件（与父节点条件AND组合）
        type: string
      match_regex:
        additionalProperties:
          type: string
        type: object
      match_regex_json:
        type: string
      match_severity:
        type: string
      name:
        description: 节点名称与标识
        type: string
      notify_resolved:
        type: boolean
      parent:
        $ref: '#/definitions/model.AlertSubscriptionNode'
      parent_id:
        description: 树形结构
        type: integer
      path:
        type: string
      project_id:
        type: integer
      receiver_group_ids:
        items:
          type: integer
        type: array
      receiver_group_ids_json:
        description: 通知配置
        type: string
      silence_seconds:
        type: integer
      updated_at:
        type: string
    type: object
  response.Body:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  service.AlertManagerPayload:
    properties:
      alerts:
        items:
          $ref: '#/definitions/alert.AlertManagerAlert'
        type: array
      commonAnnotations:
        additionalProperties:
          type: string
        type: object
      commonLabels:
        additionalProperties:
          type: string
        type: object
      externalURL:
        type: string
      groupLabels:
        additionalProperties:
          type: string
        type: object
      receiver:
        type: string
      status:
        type: string
      truncatedAlerts:
        type: integer
      version:
        type: string
    type: object
  service.AlertSubscriptionNodeUpsertRequest:
    properties:
      code:
        maxLength: 64
        type: string
      continue:
        type: boolean
      enabled:
        type: boolean
      match_labels_json:
        type: string
      match_regex_json:
        type: string
      match_severity:
        maxLength: 128
        type: string
      name:
        maxLength: 128
        type: string
      notify_resolved:
        type: boolean
      parent_id:
        type: integer
      project_id:
        type: integer
      receiver_group_ids_json:
        type: string
      silence_seconds:
        type: integer
    required:
    - name
    - project_id
    type: object
...
info:
  contact: {}
  description: |-
    YunShu CMDB is an operations CMDB console built with Gin, MySQL, Redis, SMTP mail delivery, Casbin, Cobra and slog.
    Request an email verification code first, then login with email code or register a new account by email. Username/password login remains available as an admin fallback.
  title: YunShu CMDB API
  version: "1.0"
paths:
...
  /api/v1/alerts/subscriptions:
    get:
      consumes:
      - application/json
      description: 分页查询订阅树节点，支持按项目、父节点、关键词过滤
      parameters:
      - description: 项目ID
        in: query
        name: project_id
        required: true
        type: integer
      - description: 父节点ID，0表示根节点
        in: query
        name: parent_id
        type: integer
      - description: 关键词
        in: query
        name: keyword
        type: string
      - description: 是否启用
        in: query
        name: enabled
        type: boolean
      - description: 页码，默认1
        in: query
        name: page
        type: integer
      - description: 每页数量，默认20
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/response.Body'
      summary: 查询订阅树节点列表
      tags:
      - alert-subscription
    post:
      consumes:
      - application/json
      description: 创建新的订阅树节点
      parameters:
      - description: 节点信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/service.AlertSubscriptionNodeUpsertRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.AlertSubscriptionNode'
      summary: 创建订阅树节点
      tags:
      - alert-subscription
  /api/v1/alerts/subscriptions/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定订阅树节点（有子节点时无法删除）
      parameters:
      - description: 节点ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/response.Body'
      summary: 删除订阅树节点
      tags:
      - alert-subscription
    put:
      consumes:
      - application/json
      description: 更新指定订阅树节点
      parameters:
      - description: 节点ID
        in: path
        name: id
        required: true
        type: integer
      - description: 节点信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/service.AlertSubscriptionNodeUpsertRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.AlertSubscriptionNode'
      summary: 更新订阅树节点
      tags:
      - alert-subscription
  /api/v1/alerts/subscriptions/tree:
    get:
      consumes:
      - application/json
      description: 获取指定项目的完整订阅树结构（包含所有层级）
      parameters:
      - description: 项目ID
        in: query
        name: project_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.AlertSubscriptionNode'
            type: array
      summary: 获取完整订阅树
      tags:
      - alert-subscription
  /api/v1/alerts/webhook/alertmanager:
    post:
      consumes:
      - application/json
      description: Ingest Alertmanager notifications. Auth via header X-Alert-Token
        or Authorization Bearer (not query token).
      parameters:
      - description: Webhook token (preferred)
        in: header
        name: X-Alert-Token
        type: string
      - description: Alertmanager payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/service.AlertManagerPayload'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/response.Body'
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/response.Body'
        "401":
          description: invalid webhook token
          schema:
            $ref: '#/definitions/response.Body'
      summary: Receive Alertmanager webhook
      tags:
      - AlertsWebhook
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Enter the JWT token with the `Bearer ` prefix, for example: Bearer
      eyJhbGciOi...'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"